requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
get==2019.4.13
idna==2.8
lxml==4.3.4
post==2019.4.13
psycopg2-binary==2.8.3
public==2019.4.13
py-postgresql==1.2.1
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
soupsieve==1.9.2
urllib3==1.25.3
